Avoid cloning too early
It may be necessary to pass current state of job to next attempt,
if current attempt fails.

This may be helpful in situations where a job is long-running,
and keeps track of its state internally. When the task is retried,
the job may want to continue where it left off in the previous run.