Fix concurrency bug
A task could be ran multiple times, once by each worker, if multiple
workers fetched due tasks from the sorted set before the first
worker moves the task to *pending*.

This is because reads from the sorted set and moves to *pending*
were not atomic. They could not be, because we needed the due task IDs
first before we could move to *pending*.

This commit adds a separate queue for pending tasks, so the first
worker that moves tasks to the pending queue ensures subsequent
workers do not have access to the tasks at all.